xingxianglei/mindspores/requirements.txt
imageio
matplotlib
mindspore
tqdm
scikit-image
pillow